requirements.txt
Flask==2.0.1
Pillow==8.3.1
requests==2.26.0
Werkzeug==2.2.2
gunicorn==20.1.0